INFRA-5756: add an onlyif condition to the exec

+++ manifests/packages.pp
@@ -15,6 +15,7 @@ class system::packages (
         command => "yum -y -q remove $title",
         before => Package["$title"],
         require => $require,
+        onlyif => "rpm --quiet -q $title",
       }
       package {"$title":
         ensure => absent,
